Corrigindo estoqueLivroController.js

+++ src/controller/estoqueLivroController.js
@@ -3,7 +3,6 @@ const Livro = require('../models/livroModel');
 const Estoque = require('../models/estoqueModel');
 const Tamanho = require('../models/tamanhoModel');
 
-// POST - cria um novo estoque >necessário autentificação< 
 async function adicionarEstoque(req,res){
     try{
         let livroEncontrado = null;
@@ -32,9 +31,9 @@ async function adicionarEstoque(req,res){
     }
 }
 
-// PUT/:id - edita um estoque >necessário autentificação<
 async function editarEstoque(req,res){
     const { id } = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)) return res.status(400).json({msg:"Parâmetro inválido"});
     try{
         let livroEncontrado = null;
         let tamanhoEncontrado = null;
@@ -71,13 +70,15 @@ async function editarEstoque(req,res){
     }
 }
 
-// GET - recebe todos os estoques
 async function listarEstoque(req,res){
-    const estoques = await Estoque.find({}).populate('livro').populate('tamanho');
-    return res.status(200).json(estoques);
+    try{
+        const estoques = await Estoque.find({}).populate('livro').populate('tamanho');
+        return res.status(200).json(estoques);    
+    }catch(err){
+        return res.status(500).json({ msg: "Erro interno do servidor" });
+    }
 }
 
-// GET/:id - recebe os dados do estoque e do livro 
 async function buscarEstoques(req,res,next){
     const { id } = req.params;
     if(!mongoose.Types.ObjectId.isValid(id)) return res.status(400).json({msg:"Parâmetro inválido"});
@@ -95,11 +96,16 @@ async function exibirEstoque(req,res){
     return res.status(200).json(req.estoque);
 }
 
-// DELETE - deleta o estoque >necessário autentificação<
 async function deletarEstoque(req,res){
     const { id } = req.params;
-    const estoqueDeletado = await Estoque.findOneAndDelete({_id:id});
-    return res.status(204).json({});
+    if(!mongoose.Types.ObjectId.isValid(id)) return res.status(400).json({msg:"Parâmetro inválido"});
+
+    try{
+        const estoqueDeletado = await Estoque.findOneAndDelete({_id:id});
+        return res.status(204).json({});    
+    }catch(err){
+        return res.status(500).json({ msg: "Erro interno do servidor" });
+    }
 }
 
 module.exports = { adicionarEstoque, editarEstoque, listarEstoque, buscarEstoques, exibirEstoque, deletarEstoque};
